+++ src/main/frontend/src/components/itemPage/ItemOverview.js
@@ -11,7 +11,7 @@ import { AuthContext } from '../loginAndRegistration/AuthContext';
 import BiddersSection from './BiddersSection';
 import { Link } from "react-router-dom";
 import { placeBid } from '../landingPage/ItemService';
-import { TimeInterval } from 'time-interval-js';
+import { calculateTimeLeft } from "../Utils";
 
 function ItemOverview({...item}) {
 
@@ -41,7 +41,7 @@ function ItemOverview({...item}) {
     useEffect(() => {
         setHighestBid(calculateHighestBid());
         setNoOfBids(calculateNumberOfBids());
-        setTimeLeft(calculateTimeLeft());
+        setTimeLeft(calculateTimeLeft(endDate));
     }, [itemBids]);
 
     useEffect(() => {
@@ -81,20 +81,6 @@ function ItemOverview({...item}) {
             return sortedbids[sortedbids.length-1].amount;
         } else {
             return 0;
-        }
-    }
-
-    const calculateTimeLeft = () => {
-        const date1 = new Date(endDate);
-        const date2 = new Date();
-        const interval = TimeInterval.fromTimeBetweenTwoDates(date1, date2);
-        const hours = interval.inHours();
-        if (hours < 24) {
-            return Math.round(hours) + " hours";
-        } else if (hours >= 24 && hours < 168) {
-            return Math.round(hours/24) + " days " + Math.round(hours%24) + " hours" ;
-        } else {
-            return Math.round(hours/186) + " weeks " + Math.round((hours%168)/24) + " days";
         }
     }
