+++ src/main/java/com/example/auctionapp/service/ItemService.java
@@ -32,8 +32,7 @@ public class ItemService {
 
     public Page<Item> getItems(int page, int size, ItemSort sort, Sort.Direction direction) {
         PageRequest pageable = PageRequest.of(page, size, Sort.by(direction, sort.toString()));
-        Page<Item> items = itemRepository.findAll(pageable);
-        return items;
+        return itemRepository.getItems(pageable);
     }
 
     public Optional<Item> getItemById(long id) {
